Moved room actions into MultiUserChat
